Fix date bug
* There was a bug with the date fields in Task objects in that they
  were converting to the wrong date. This was because the date in
  milliseconds is too large for an int.
* All date fields and references have been converted from int to long.
* This was a problem on the database end too, because INTEGERs in
  a SQLite database can only be 8 bits (too small for long int).
* This was fixed by dividing the millisecond value by 60000 to convert
  it to minutes. This allows the data to be small enough to be
  reliably stored.
* A downside is that the precision of the date fields have been reduced
  to minutes, but this isn't really a downside because for the app's
  purposes that is precise enough.

+++ src/edu/worcester/cs499summer2012/comparator/TaskAutoComparator.java
@@ -82,7 +82,7 @@ public class TaskAutoComparator implements Comparator<Task> {
 					// Ex. LHS -> date 5000ms (earlier)
 					//     RHS -> date 6000ms (later)
 					//     LHS - RHS = 5000 - 6000 = -1000 = LHS ordered first
-					return lhs.getDateDue() - rhs.getDateDue();
+					return (int) (lhs.getDateDue() - rhs.getDateDue());
 				}
 			}
 		}
