+++ src/edu/worcester/cs499summer2012/comparator/TaskFinalDateDueComparator.java
@@ -60,7 +60,7 @@ public class TaskFinalDateDueComparator implements Comparator<Task> {
 				// Ex. LHS -> date 5000ms (earlier)
 				//     RHS -> date 6000ms (later)
 				//     LHS - RHS = 5000 - 6000 = -1000 = LHS ordered first
-				return lhs.getFinalDateDue() - rhs.getFinalDateDue();
+				return (int) (lhs.getFinalDateDue() - rhs.getFinalDateDue());
 			}
 		}
 	}
